+++ PRACTICA 2/plantillaSDL/HolaSDL/Player.h
@@ -41,7 +41,7 @@ private:
 
 	int walkFrame;
 
-	Point2D<int> initPos;
+	Point2D<double> initPos;
 
 };
